Moved logic from TachiyomiController.cs to TachiyomiService.cs
